requirements.txt
requests
Pillow
pyyaml
